added loads of new toys to the bezier curve tools.

+++ Drydock/Drydock/Logic/HullEditorState/BezierCurveCollection.cs
@@ -298,7 +298,7 @@ namespace Drydock.Logic.HullEditorState{
                 prevCurve.NextHandlePos,
                 nextCurve.PrevHandlePos,
                 nextCurve.CenterHandlePos,
-                t
+                (float)t
                 );
 
             return retVal;
